+++ src/game-log/game.module.ts
@@ -4,12 +4,13 @@ import { getRepositoryToken, TypeOrmModule } from '@nestjs/typeorm';
 import { AppModule } from 'src/app.module';
 import { GameGateway } from './game.gateway';
 import { GameService } from './game.service';
-import { GameLogRepository, MockGameLogRepository } from './game-log.repository';
+import { GameLogRepository } from './repository/game-log.repository';
 import { GameQueue } from './game-queue';
 import { SimulationService } from './simulation.service';
 import { GameSocketSession } from './game-socket-session';
 import { GameLogController } from './game-log.controller';
 import { GameLogService } from './game-log.service';
+import { MockGameLogRepository } from './repository/mock/mock.game-log.repository';
 
 @Module({
   imports: [
